+++ packages/d2mini/src/graph.ts
@@ -115,3 +115,16 @@ export abstract class BinaryOperator<T> extends Operator<T> {
     return this.inputs[1].drain()
   }
 }
+
+/**
+ * Base class for operators that process a single input stream
+ */
+export abstract class LinearUnaryOperator<T, U> extends UnaryOperator<T | U> {
+  abstract inner(collection: MultiSet<T | U>): MultiSet<U>
+
+  run(): void {
+    for (const message of this.inputMessages()) {
+      this.output.sendData(this.inner(message))
+    }
+  }
+}
