+++ packages/d2mini/src/index.ts
@@ -1,5 +1,4 @@
 export * from './d2.js'
 export * from './multiset.js'
-// export * from './version-index.js'
 export * from './operators/index.js'
 export * from './types.js'
